requirements_test.txt
#test/debuggingtools
#django-test-plus
#Werkzeug
django_extensions
attrs
coverage
factory-boy
Faker
# iniconfig
# packaging
# pluggy
# py
# pyparsing
pytest
pytest-cov
pytest-django
# python-dateutil
# tomli
mixer